requirements.txt
flet>=0.10.0
cryptography>=40.0.0
pillow>=9.5.0
argon2-cffi>=23.1.0
python-dotenv>=1.0.0
fastapi
uvicorn
httpx
websockets
playsound>=1.3.0 